db+sqlite+pg: split GetAddress into its own file
Move GetAddress into addressstore_getaddress.go in each package. It is a thin
read (one script-pubkey lookup plus row conversion), so per the
wallet/internal/db README it stays a direct backend implementation rather than
an ops interface: each backend validates the query, runs its sqlc lookup, maps
ErrNoRows to ErrAddressNotFound, and converts the row.
